build - move info and warning to buildlib to allow use from makers
